Fix: User와 Schedule 사이의 연관관계 설정으로 인한 ScheduleService 수정
- User와 Schedule 사이의 연관관계 설정으로 인한, UserRepository 의존성 주입
- ScheduleRequestDto 병합
- 해당 schedule id의 일정 조회 메서드명 변경
- 불필요한 주석 삭제

+++ src/main/java/com/eun0/schedulerdevelop/service/ScheduleService.java
@@ -1,11 +1,13 @@
 package com.eun0.schedulerdevelop.service;
 
-import com.eun0.schedulerdevelop.dto.schedule.ScheduleCreateRequest;
 import com.eun0.schedulerdevelop.dto.schedule.SchedulePagingResponse;
+import com.eun0.schedulerdevelop.dto.schedule.ScheduleRequest;
 import com.eun0.schedulerdevelop.dto.schedule.ScheduleResponse;
-import com.eun0.schedulerdevelop.dto.schedule.ScheduleUpdateRequest;
 import com.eun0.schedulerdevelop.entity.Schedule;
+import com.eun0.schedulerdevelop.entity.User;
 import com.eun0.schedulerdevelop.repository.ScheduleRepository;
+import com.eun0.schedulerdevelop.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -13,31 +15,24 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@RequiredArgsConstructor
 public class ScheduleService {
+    UserRepository userRepository;
     ScheduleRepository scheduleRepository;
 
-    public ScheduleService(ScheduleRepository scheduleRepository) {
-        this.scheduleRepository = scheduleRepository;
-    }
-
     @Transactional
-    public ScheduleResponse createSchedule(ScheduleCreateRequest requestDto) {
-        // RequestDTO -> Entity
-        Schedule schedule = requestDto.toEntity();
+    public ScheduleResponse createSchedule(ScheduleRequest requestDto, Long userId) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new IllegalArgumentException("해당 유저가 DB에 존재하지 않습니다."));
 
-        // DB 저장
-        Schedule savedSchedule = scheduleRepository.save(schedule);
+        Schedule savedSchedule = scheduleRepository.save(Schedule.from(requestDto, user));
 
-        // Entity -> ResponseDTO
         return ScheduleResponse.from(savedSchedule);
     }
 
     @Transactional
-    public ScheduleResponse readScheduleById(Long id) {
-        // 해당 일정이 DB에 존재하는지 확인
+    public ScheduleResponse readSchedule(Long id) {
         Schedule readSchedule = scheduleRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 일정이 DB에 존재하지 않습니다."));
 
-        // Entity -> ResponseDTO
         return ScheduleResponse.from(readSchedule);
     }
 
@@ -49,21 +44,26 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleResponse updateSchedule(Long id, ScheduleUpdateRequest requestDto) {
-        // 해당 일정이 DB에 존재하는지 확인
+    public ScheduleResponse updateSchedule(Long id, ScheduleRequest requestDto, Long userId) {
         Schedule schedule = scheduleRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 스케줄이 DB에 존재하지 않습니다."));
 
-        // RequestDTO -> Entity
+        if (userId.equals(schedule.getUser().getId())) {
+            throw new IllegalArgumentException("해당 일정의 작성자가 아닙니다.");
+        }
+
         schedule.update(requestDto.getTitle(), requestDto.getContent());
 
-        // Entity -> ResponseDTO
         return ScheduleResponse.from(schedule);
     }
 
+
     @Transactional
-    public void deleteSchedule(Long id) {
-        // 해당 일정이 DB에 존재하는지 확인
+    public void deleteSchedule(Long id, Long userId) {
         Schedule schedule = scheduleRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 스케줄이 DB에 존재하지 않습니다."));
+
+        if (userId.equals(schedule.getUser().getId())) {
+            throw new IllegalArgumentException("해당 일정의 작성자가 아닙니다.");
+        }
 
         scheduleRepository.delete(schedule);
     }
